+++ lib/src/tools/circle_data.dart
@@ -1,5 +1,18 @@
 import 'dart:ui';
 
+/// A class representing data for a circle, including its unique identifier,
+/// normalized position, radius, and color.
+///
+/// The `normalizedPosition` represents the position of the circle as a
+/// normalized [Offset] where both `dx` and `dy` values must be between 0 and 1
+/// (inclusive). This ensures the position is within a normalized coordinate
+/// space.
+///
+/// The `radius` defines the size of the circle, and the `color` specifies its
+/// visual appearance.
+///
+/// Throws an [AssertionError] if the `normalizedPosition` values are not
+/// within the range of 0 to 1.
 class CircleData {
   final String id;
   final Offset normalizedPosition;