+++ lib/transition_type.dart
@@ -1,3 +1,14 @@
+/// An enumeration that defines various types of transitions that can be used
+/// for animations or navigation effects.
+///
+/// The available transition types are:
+/// - `leftToRight`: Transition from left to right.
+/// - `rightToLeft`: Transition from right to left.
+/// - `topToBottom`: Transition from top to bottom.
+/// - `bottomToTop`: Transition from bottom to top.
+/// - `fade`: A fade-in or fade-out transition.
+/// - `scale`: A scaling transition, where the element grows or shrinks.
+/// - `rotate`: A rotational transition.
 enum TransitionType {
   leftToRight,
   rightToLeft,